api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "afonso.medeiros"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bottle (>=0.13.2,<0.14.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
